requirements.txt
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.0.1
reportlab==4.0.5
scikit-learn==1.3.1
nltk==3.8.1
transformers==4.35.0
torch>=2.2.0                                                                                       
